Factor out `IssueList` from `VerifyScreen`
We'll use this to filter issues, and so it will no longer be local to this class.
